# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Local microservices development environment API",
        "title": "DevArch API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8550",
    "basePath": "/api/v1",
    "paths": {
...
        "/stacks/{name}/restart": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Restart all instances in a stack. Stack must be enabled.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stacks"
                ],
                "summary": "Restart a stack",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stack name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ActionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/stacks/{name}/start": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Start all instances in a stack. Stack must be enabled.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stacks"
                ],
                "summary": "Start a stack",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stack name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ActionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/stacks/{name}/stop": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stacks"
                ],
                "summary": "Stop a stack",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stack name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ActionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_priz_devarch-api_internal_api_respond.ActionResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "output": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_priz_devarch-api_internal_api_respond.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorDetail"
                }
            }
        },
        "github_com_priz_devarch-api_internal_api_respond.SuccessEnvelope": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key authentication header",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}